+++ src/managers/actuator_manager.h
@@ -2,6 +2,6 @@
 
 namespace actuator_manager {
     void initialize();
-    void open_valve_by_percent(unsigned int opening_percentage);
+    void open_valve_by_percent(float opening_percentage);
     void close_valve();
 }
